Peer Code Review
Made some changes to your else if statements in order to take care of
some redundancy.  Added a textbox render() method to show you how to use
a textbox for you future projects.  If you create a surface for your
program, the textbox and render method make showing the user a textbox
much more efficient than using drawString a bunch of times.  I commented
out all of the render methods but I also showed you where to comment and
uncomment in order to use render.